Add arena things
